adding the current database file , compeleteing functions for room and updating the access function so could be used in routes

+++ App/Routers/routes.php
@@ -1,5 +1,6 @@
 <?php
 use App\Controllers\DestinationController;
+use App\Controllers\RoomController;
 use App\Routers\Router as Router;
 use App\Middlewares\AuthMiddleware;
 
@@ -24,14 +25,17 @@ $router->post('v1','/login', AuthController::class, 'login');
 $router->post('v1','/register', AuthController::class, 'register');
 $router->post('v1','/verify', AuthController::class, 'verify');
 // weathers
-$router->get('v1', '/weathers', WeatherController::class, 'index');
-$router->get('v1', '/weathers/{id}', WeatherController::class, 'get');
-$router->put('v1', '/weathers/{id}', WeatherController::class, 'update');
-$router->delete('v1', '/weathers/{id}', WeatherController::class, 'destroy');
-$router->post('v1', '/weathers', WeatherController::class, 'store');
+$router->get('v1', '/weathers', WeatherController::class, 'index',['support','admin']);
+$router->get('v1', '/weathers/{id}', WeatherController::class, 'get',['support','admin']);
+$router->put('v1', '/weathers/{id}', WeatherController::class, 'update',['support','admin']);
+$router->delete('v1', '/weathers/{id}', WeatherController::class, 'destroy',['support','admin']);
+$router->post('v1', '/weathers', WeatherController::class, 'store',['support','admin']);
 // Destentions
-$router->get('v1', '/destinations', DestinationController::class, 'index');
-$router->get('v1', '/destinations/{id}', DestinationController::class, 'get');
-$router->post('v1', '/destinations', DestinationController::class, 'store');
-$router->put('v1', '/destinations/{id}', DestinationController::class, 'update');
+$router->get('v1', '/destinations', DestinationController::class, 'index',['support','admin']);
+$router->get('v1', '/destinations/{id}', DestinationController::class, 'get',['support','admin']);
+$router->post('v1', '/destinations', DestinationController::class, 'store',['support','admin']);
+$router->put('v1', '/destinations/{id}', DestinationController::class, 'update',['support','admin']);
 $router->delete('v1', '/destinations/{id}', DestinationController::class, 'destroy');
+// Rooms
+$router->get('v1', '/rooms', RoomController::class, 'index',['host','support','admin']);
+$router->post('v1', '/rooms', RoomController::class, 'store',['host','support','admin']);
